+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2022-02-13
+
 ### Added
 
 -   Add support for Microsoft Store version of Hades.
@@ -302,7 +304,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Initial release.
 
-[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v1.5.1...HEAD
+[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v1.6.0...HEAD
+
+[1.6.0]: https://github.com/nbusseneau/hephaistos/compare/v1.5.1...v1.6.0
 
 [1.5.1]: https://github.com/nbusseneau/hephaistos/compare/v1.5.0...v1.5.1